+++ examples/app.rs
@@ -1,3 +1,6 @@
+use std::path::PathBuf;
+use std::ffi::OsStr;
+
 use pico_args::Arguments;
 
 #[derive(Debug)]
@@ -6,11 +9,16 @@ struct AppArgs {
     number: u32,
     opt_number: Option<u32>,
     width: u32,
+    input: Option<PathBuf>,
     free: Vec<String>,
 }
 
 fn parse_width(s: &str) -> Result<u32, &'static str> {
     s.parse().map_err(|_| "not a number")
+}
+
+fn parse_path(s: &OsStr) -> Result<PathBuf, &'static str> {
+    Ok(s.into())
 }
 
 fn main() {
@@ -20,12 +28,13 @@ fn main() {
 }
 
 fn submain() -> Result<(), pico_args::Error> {
-    let mut args = Arguments::from_env()?;
+    let mut args = Arguments::from_env();
     let args = AppArgs {
         help: args.contains(["-h", "--help"]),
         number: args.value_from_str("--number")?.unwrap_or(5),
         opt_number: args.value_from_str("--opt-number")?,
         width: args.value_from_fn("--width", parse_width)?.unwrap_or(10),
+        input: args.value_from_os_str("--input", parse_path)?,
         free: args.free()?,
     };